fixed mistakes

+++ src/main/resources/init_db.sql
@@ -1,9 +1,10 @@
 CREATE DATABASE books;
 USE books;
 CREATE TABLE books(
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
     price DECIMAL,
     PRIMARY KEY(id)
 );
+
